stock-minority-report/requirements.txt
streamlit==1.45.1
pandas==2.2.3
numpy==2.2.6
plotly==6.0.1
yfinance==0.2.58
python-dotenv==1.0.1
openai==1.79.0
anthropic==0.54.0
google-generativeai==0.8.5
pytest==8.3.5
ruff==0.11.10
